Add tests. Move components.
