+++ app/components/__tests__/Calendar.spec.tsx
@@ -1,4 +1,4 @@
-import Calendar from '../layout/Calendar';
+import Calendar from '../Calendar';
 import { Provider } from 'react-redux';
 import { makeStore } from '../../redux/store';
 import renderer from 'react-test-renderer';
@@ -6,13 +6,21 @@ import renderer from 'react-test-renderer';
 describe('<Calendar />', () => {
   it('renders correctly', () => {
     const store = makeStore();
+
+    const mockCalendarChange = (): void => {
+      true;
+    };
+
+    const mockScrollPosition = (): void => {
+      true;
+    };
+
     const tree = renderer
       .create(
         <Provider store={store}>
           <Calendar
-            handleCalendarChange={() => {
-              true;
-            }}
+            handleCalendarChange={mockCalendarChange}
+            handleScrollPosition={mockScrollPosition}
           />
         </Provider>
       )
