+++ app/components/__tests__/DayItem.spec.tsx
@@ -1,4 +1,4 @@
-import DayItem from '../layout/DayItem';
+import DayItem from '../DayItem';
 import { Provider } from 'react-redux';
 import { makeStore } from '../../redux/store';
 import renderer from 'react-test-renderer';
