+++ app/components/__tests__/Tradtion.spec.tsx
@@ -1,20 +1,15 @@
-import Calendar from '../layout/Calendar';
 import { Provider } from 'react-redux';
+import Tradition from '../Tradition';
 import { makeStore } from '../../redux/store';
 import renderer from 'react-test-renderer';
 
-describe('<Calendar />', () => {
+describe('<Tradition />', () => {
   it('renders correctly', () => {
     const store = makeStore();
-
-    const mockFunction = (): void => {
-      true;
-    };
-
     const tree = renderer
       .create(
         <Provider store={store}>
-          <Calendar handleCalendarChange={mockFunction} />
+          <Tradition id={''} tradition={''} />
         </Provider>
       )
       .toJSON();
