+++ app/screens/__tests__/About.spec.tsx
@@ -1,0 +1,9 @@
+import About from '../About';
+import renderer from 'react-test-renderer';
+
+describe('<About />', () => {
+  it('renders correctly', () => {
+    const tree = renderer.create(<About />).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
